requirements.txt
annotated-types==0.5.0
catboost==1.2
certifi==2023.5.7
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
fonttools==4.41.0
graphviz==0.20.1
idna==3.4
joblib==1.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
plotly==5.15.0
pydantic==2.0.3
pydantic_core==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.1
six==1.16.0
tenacity==8.2.2
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
